refactor(oauth): make token_exchange private
Establishes the rule the package now follows: a module is underscore-prefixed
exactly when nothing outside bfabric.oauth imports it. token_exchange is the
only one that qualifies — exchange_token has a single caller (webapp_client)
and introspect_token has none.

The login modules keep plain names deliberately: bfabric_scripts is a separate
distribution and drives `auth login` / `auth status` through pkce,
device_code, credential_provider, token_cache and registration, so hiding them
would only move the private-import problem across the package boundary.

url_token stays plain for UrlTokenContext even though its verify_jwt is
internal.
