+++ tests/bfabric/oauth/test_webapp_client.py
@@ -7,7 +7,7 @@ import pytest
 from bfabric.oauth.url_token import UrlTokenContext
 from bfabric.oauth.webapp_client import WebappClient
 
-_PATCH_EXCHANGE = "bfabric.oauth.token_exchange.exchange_token"
+_PATCH_EXCHANGE = "bfabric.oauth._token_exchange.exchange_token"
 _PATCH_VERIFY_JWT = "bfabric.oauth.url_token.verify_jwt"
 _PATCH_PROVIDER = "bfabric.oauth.credential_provider.OAuthCredentialProvider"
 _PATCH_CONNECT_OAUTH = "bfabric.bfabric.Bfabric.connect_oauth"
